% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.r
\name{add_snippets}
\alias{add_snippets}
\title{Add qcscript Snippets to Rstudio}
\usage{
add_snippets()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snipped definitions were added
}
\description{
\code{add_snippets} copies all (missing) snippet definitions
  in 'inst/Rsnippets.txt' to the RStudios user snippet location.
}
\examples{
\dontrun{add_snippets()}
}
