% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSnippets.r
\name{addQCscript}
\alias{addQCscript}
\title{Add qcscript Snippets to Rstudio}
\usage{
addQCscript()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snipped definitions were added
}
\description{
\code{addQCscript} copies all (missing) snippet definitions
  in 'inst/Rsnippets.txt' to the RStudios user snippet location.
}
\examples{
\dontrun{addPackageSnippets()}
}
