% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{cat_row}
\alias{cat_row}
\title{Create Rows for Categorical Variables}
\usage{
cat_row(data, colvar, rowvar, N_row, keep = TRUE)
}
\arguments{
\item{data}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable}

\item{N_row}{a dataframe with denominator N}

\item{keep}{if = TRUE, keep all factor levels}
}
\value{
dataframe with demographic rows
}
\description{
Function to create rows for categorical variables in demographic table
}
\examples{
cat_row(adsl, "TRT01P", "SEX", N_row = firstrow)
}
