% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfchg.r
\name{chgfb}
\alias{chgfb}
\title{Create change from basline over time table}
\usage{
chgfb(
  input,
  val,
  chg,
  rowvar,
  stats_list = c("N", "Mean", "SD", "Median", "Min", "Max"),
  max_digit = 2,
  keep = TRUE
)
}
\arguments{
\item{input}{input dataframe}

\item{val}{measurement variable}

\item{chg}{change from baseline measurement variable}

\item{rowvar}{row variable (can be a set of two variables such as c("TRT01P", "AVISIT") or three variables such as c("PARAM", "TRT01P", "AVISIT"))}

\item{stats_list}{stats variables to display. Accepted values are c("N", "Mean", "SD", "Median", "Min", "Max", "CV", "Base_mean") and the order of variables matters}

\item{max_digit}{largest number of digit to display}

\item{keep}{if = TRUE, keep all levels. Default = FALSE}
}
\value{
change from baseline table
}
\description{
create change from basline over time table, table can be ordered by param, treatment group and visit
}
\examples{
chgfb(adlb, "AVAL", "CHG", rowvar = c("PARAM", "TRT01P", "AVISIT"), max_digit = 3, keep = FALSE)
}
