% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountpct.r
\name{cntpct}
\alias{cntpct}
\title{Compute count and percentage}
\usage{
cntpct(
  input,
  colvar = "TRT01P",
  row_text = "Subjects with 1 or more AEs",
  N_row,
  subset = NULL
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{row_text}{row text}

\item{N_row}{dataframe with N}

\item{subset}{subset criteria. Default = NULL}
}
\value{
a dataframe containing count and percentage by colvar
}
\description{
compute count and percentage by column variable
}
