% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountpctpt.r
\name{cntpctpt}
\alias{cntpctpt}
\title{Compute count and percentage of SOC/PT}
\usage{
cntpctpt(
  input,
  colvar = "TRT01P",
  rowvar = c("AEBODSYS", "AEDECOD"),
  row_text = "Subjects with 1 or more AEs",
  N_row,
  col_order = NULL,
  subset = NULL
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable (can be single variable such as "AEDECOD" or multiple variable such as c("AEBODSYS", "AEDECOD"))}

\item{row_text}{row text}

\item{N_row}{dataframe with N}

\item{col_order}{ordering column name. e.g. "n_5". Default is NULL and PTs will be ordered by row-wise sum}

\item{subset}{subset criteria. Default = NULL}
}
\value{
a dataframe containing count and percentage of SOC/PT by colvar
}
\description{
compute count and percentage of SOC/PT by column variable with an ordering
}
\examples{
countpctpt(qc = adae, colvar = "TRT01PN", rowvar = "AEDECOD",
           row_text = "Subjects with 1 or more AEs", 
}
