% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountrow1.r
\name{countrow1}
\alias{countrow1}
\title{Create Analysis Set Row}
\usage{
countrow1(
  input,
  colvar = "TRT01P",
  row_text = "Analysis set: Safety",
  subset = "TRUE"
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{row_text}{row text}

\item{subset}{subset criteria. Default = "TRUE" means no subsetting}
}
\value{
Analysis set row
}
\description{
Function to create analysis set row
}
\examples{
cntrow1(input = adae, colvar = "TRT01P", row_text = "Analysis set: Safety")
}
