% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{demo}
\alias{demo}
\title{Create Demographic Rows}
\usage{
demo(
  input,
  colvar = "TRT01P",
  N_row,
  stats_list = c("Mean_SD", "Median", "Range"),
  var_list = c("AGE", "AGEGR1", "SEX"),
  con_var_list = "AGEGR1",
  drop_var_list = NULL,
  max_digit = 2
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{N_row}{a dataframe with denominator N}

\item{stats_list}{stats variables to display. Accepted values are c("Mean_SD", "Median", "Range", "Geo_mean", "Geo_CV", "Geo_CL") and the order of variables matters}

\item{var_list}{variable list in demo table}

\item{con_var_list}{List of continuous variables that concatenate to the corresponding categorical variables}

\item{drop_var_list}{Variable that don't keep all levels in the output}

\item{max_digit}{maximum number of decimal place to report}
}
\value{
dataframe with demographic rows
}
\description{
Function to create demographic rows
}
\examples{
### Create variable list based on DPS
var_list <- c("AGE", "AGEGR1", "SEX", "RACEGR1", "ETHNIC", "REGION2", "REGION1", "WEIGHTBL", "HEIGHTBL",
             "BMIBL", "BMIBLG1", "WHOFCBL", "PAHETBL", "PAHDURY")
             
# list of variable labels displayed in the table
names(var_list) <- c("Age, years", "", "Sex", "Race", "Ethnicity", "Region", "Geographical region", "Weight, kg", "Height, cm", 
                    "Body mass index, kg/m2", "", "WHO FC", " PAH etiology", "Time since diagnosis, years")
                    
# List of continuous variables that concatenate to the corresponding categorical variables
var_list1 <- c("AGEGR1", "BMIBLG1")

# Variable that don't keep all levels in the output
var_list2 <- "PAHETBL"

tab_final <- demo(adsl1, colvar = "TRT01PN", N_row = first_row[[1]], var_list = var_list, 
                 con_var_list = var_list1, drop_var_list = var_list2)
}
