% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{num_row}
\alias{num_row}
\title{Create Rows for Continuous Variables}
\usage{
num_row(data, colvar = "TRT01P", rowvar = "AGE", stats_list, digit)
}
\arguments{
\item{data}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable}

\item{stats_list}{stats variables to display. Accepted values are c("Mean_SD", "Median", "Range", "Geo_mean", "Geo_CV", "Geo_CL") and the order of variables matters}

\item{digit}{number of decimal place to report}
}
\value{
dataframe with demographic rows
}
\description{
Function to create rows for continuous variables in demographic table
}
\examples{
num_row(adsl, "TRT01P", "AGE", digit = 1)
}
