% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{qc_cat_row}
\alias{qc_cat_row}
\title{Create Rows for Categorical Variables}
\usage{
qc_cat_row(
  input,
  colvar = "TRT01P",
  rowvar = "SEX",
  row_text = "Sex",
  keep = TRUE
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable}

\item{row_text}{row text}

\item{keep}{if = TRUE, keep all factor levels}
}
\value{
dataframe with demographic rows
}
\description{
Function to create rows for categorical variables in demographic table
}
\examples{
adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = factor(sample(c("A", "B", "C"), 10, replace = TRUE)),
  SEX = factor(sample(c("Female", "Male"), 10, replace = TRUE)))
  
tab1 <- qc_cat_row(adsl, "TRT01P", "SEX")
tab1
}
