% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountpct.r
\name{qc_cntpct}
\alias{qc_cntpct}
\title{Create Count and Percentage Row}
\usage{
qc_cntpct(
  input,
  colvar = "TRT01P",
  row_text = "Subjects with 1 or more AEs",
  N_row,
  subset = NULL
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{row_text}{row text}

\item{N_row}{dataframe with N}

\item{subset}{subset criteria. Default = NULL}
}
\value{
a dataframe containing count and percentage by colvar
}
\description{
compute count and percentage by column variable
}
\examples{
aedecod <- sample(paste0("PT", 1:3), 10, replace = TRUE)

adae <- data.frame(
  USUBJID = 1:10,
  TRT01P = factor(sample(c("A", "B", "C"), 10, replace = TRUE)),
  SEX = factor(sample(c("Female", "Male"), 10, replace = TRUE)),
  AEBODSYS = ifelse(aedecod == "PT1", "SOC1", "SOC2"),
  AEDECOD = aedecod)

### Create analysis row first
first_row <- qc_cntrow1(input = adae, colvar = "TRT01P", row_text = "Analysis set: Safety")

tab1 <- qc_cntpct(input = adae, colvar = "TRT01P", row_text = "Subjects with 1 or more AEs", 
                  N_row = first_row$N_row)
tab1
}
