% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_cntpct_byrowvar.r
\name{qc_cntpct_byrowvar}
\alias{qc_cntpct_byrowvar}
\title{Create Count and Percentage Table by Row Variables}
\usage{
qc_cntpct_byrowvar(
  input,
  colvar = "TRT01P",
  rowvar = c("AEBODSYS", "AEDECOD"),
  row_text = "Subjects with 1 or more AEs",
  N_row,
  col_order = NULL,
  subset = NULL
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable (can be single variable such as "AEDECOD" or multiple variable such as c("AEBODSYS", "AEDECOD"))}

\item{row_text}{row text}

\item{N_row}{dataframe with N}

\item{col_order}{ordering column name. e.g. "n_5". Default is NULL and PTs will be ordered by row-wise sum}

\item{subset}{subset criteria. Default = NULL}
}
\value{
a dataframe containing count and percentage of SOC/PT by colvar
}
\description{
compute count and percentage of SOC(class)/PT by column variable with an ordering
}
\examples{
aedecod <- sample(paste0("PT", 1:3), 10, replace = TRUE)

adae <- data.frame(
  USUBJID = 1:10,
  TRT01P = factor(sample(c("A", "B", "C"), 10, replace = TRUE)),
  SEX = as.factor(sample(c("Female", "Male"), 10, replace = TRUE)),
  AEBODSYS = ifelse(aedecod == "PT1", "SOC1", "SOC2"),
  AEDECOD = aedecod)

### Create analysis row first
first_row <- qc_cntrow1(input = adae, colvar = "TRT01P", row_text = "Analysis set: Safety")

tab1 <- qc_cntpct_byrowvar(input = adae, colvar = "TRT01P", rowvar = c("AEBODSYS", "AEDECOD"),
           row_text = "Subjects with 1 or more AEs", N_row = first_row$N_row)
tab1
}
