% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountrow1_shift.r
\name{qc_cntrow1_shift}
\alias{qc_cntrow1_shift}
\title{Create Analysis Set Row for Shift Table}
\usage{
qc_cntrow1_shift(
  input,
  colvar = "TRT01P",
  row_text = "Analysis set: Full",
  subset = "TRUE"
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{row_text}{row text}

\item{subset}{subset criteria. Default = "TRUE" means no subsetting}
}
\value{
Analysis set row for shift table
}
\description{
Function to create analysis set row for shift table
}
\examples{
adlb <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = T))
  
first_row <- qc_cntrow1_shift(input = adlb, colvar = "TRT01P", row_text = "Analysis set: Full")
first_row
}
