% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_compare2xlsx.r
\name{qc_compare2xlsx}
\alias{qc_compare2xlsx}
\title{Compare QC and RTF Read-in Data Frames}
\usage{
qc_compare2xlsx(
  qc,
  rtf,
  path = NULL,
  filename = NULL,
  max_diff = 50,
  max_diff_per_var = 10,
  version = TRUE
)
}
\arguments{
\item{qc}{the name of data frame you create in QC}

\item{rtf}{the name of data frame read in from RTF}

\item{path}{the path to save XLSX output file folder}

\item{filename}{output XLSX file name, which is typically tableid}

\item{max_diff}{maximum number of differences to report. NA will print all differences overall with the constraint of max_diff_per_var}

\item{max_diff_per_var}{maximum number of differences per variable to report. NA will print all differences for each variable with the constraint of max_diff}

\item{version}{whether to add time stamp in file name. Default is TRUE to add the current date in file name}
}
\value{
the output of comparison in XLSX
}
\description{
Tool used to compare QC data frame with RTF read-in data frame, and output comparison results to XLSX file
}
\examples{
\dontrun{
qc_compare2xlsx(qc = tab_qc, rtf = tab_rtf, path = qc, filename = "tableid")
}
}
