% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_comparedf.r
\name{qc_comparedf}
\alias{qc_comparedf}
\title{Compare QC and RTF Read-in Data Frames}
\usage{
qc_comparedf(
  qc,
  rtf,
  path = ".",
  filename = NULL,
  by = "row_seq",
  exclude = NULL,
  tolerance = 0,
  tolerance_type = "ratio",
  stop_on_error = FALSE,
  keep_unchanged_rows = FALSE,
  keep_unchanged_cols = TRUE,
  round_output_to = 3,
  limit = 1000
)
}
\arguments{
\item{qc}{the name of data frame you create in QC}

\item{rtf}{the name of data frame read in from RTF}

\item{path}{the folder path of HTML output file. By default the path is the working directory, getwd().}

\item{filename}{output HTML file name, which is typically tableid}

\item{by}{A character vector of a string of character vector showing the columns by which to group_by. Defaults to row sequence}

\item{exclude}{The columns which should be excluded from the comparison}

\item{tolerance}{The amount in fraction to which changes are ignored while showing the visual representation. By default, the value is 0 and any change in the value of variables is shown off. Doesn't apply to categorical variables.}

\item{tolerance_type}{Defaults to 'ratio'. The type of comparison for numeric values, can be 'ratio' or 'difference'}

\item{stop_on_error}{Whether to stop on acceptable errors on not. Defaults to FALSE}

\item{keep_unchanged_rows}{whether to preserve unchanged values or not. Defaults to FALSE}

\item{keep_unchanged_cols}{whether to preserve unchanged values or not. Defaults to TRUE}

\item{round_output_to}{Number of digits to round the output to. Defaults to 3.}

\item{limit}{The max number of differences to display. Defaults to 1000.}
}
\value{
the output of comparison in HTML
}
\description{
Tool used to compare QC data frame with RTF read-in data frame, and output comparison results to HTML file
}
\examples{
\dontrun{
qc_comparedf(qc = tab_qc, rtf = tab_rtf, path = qc[["PDEV"]], filename = "tsidem01")
}
}
