% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{qc_demo}
\alias{qc_demo}
\title{Create Demographic Rows}
\usage{
qc_demo(
  input,
  colvar = "TRT01P",
  N_row,
  stats_list = c("Mean_SD", "Median", "Range"),
  var_list = c("AGE", "AGEGR1", "SEX"),
  con_var_list = "AGEGR1",
  drop_var_list = NULL,
  max_digit = 2
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{N_row}{a dataframe with denominator N}

\item{stats_list}{stats variables to display. Accepted values are c("Mean_SD", "Median", "Range", "Geo_mean", "Geo_CV", "Geo_CL") and the order of variables matters}

\item{var_list}{variable list in demo table. Please see details.}

\item{drop_var_list}{Variable that don't keep all levels in the output}

\item{max_digit}{maximum number of decimal place to report}
}
\value{
dataframe with demographic rows
}
\description{
Function to create demographic rows
}
\details{
var_list should be a dataframe with desired variable names as column names,
the label of each variable display in tables as the first row, and optionally,
the number of decimal places to keep in the output as the second row. If the 
label row of a variable is empty, the row text and N row for this variable
will be missing.
}
\examples{
age <- sample(18:65, 10, replace = T)

adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = T),
  AGE = age,
  AGEGR1 = ifelse(age < 45, "< 45", ">= 45"),
  SEX = as.factor(sample(c("Female", "Male"), 10, replace = T)))
  
### Create variable list based on DPS and assign labels to them.
### Leave "" For variables that concatenate to the corresponding categorical variables
var_list <- data.frame(
  AGE = "Age, years",
  AGEGR1 = "",
  SEX = "Sex"
  )

### Create analysis row first
first_row <- qc_cntrow1(input = adsl, colvar = "TRT01P", row_text = "Analysis set: Safety")

tab1 <- qc_demo(adsl, colvar = "TRT01P", N_row = N_row, var_list = var_list, 
               con_var_list = var_list1)
tab1
}
