% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{qc_num_row}
\alias{qc_num_row}
\title{Create Rows for Continuous Variables}
\usage{
qc_num_row(
  input,
  colvar = "TRT01P",
  rowvar = "AGE",
  row_text = "Age, years",
  stats_list,
  digit
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable}

\item{row_text}{row text}

\item{stats_list}{stats variables to display. Accepted values are \code{c("Mean_SD", "Median", "Range", "Geo_mean", "Geo_CV", "Geo_CL")} and the order of variables matters}

\item{digit}{number of decimal place to report}
}
\value{
dataframe with demographic rows
}
\description{
Function to create rows for continuous variables in demographic table
}
\examples{
adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = TRUE),
  AGE = sample(18:65, 10, replace = TRUE))
  
tab1 <- qc_num_row(input = adsl, colvar = "TRT01P", rowvar = "AGE", 
                        stats_list = c("Mean_SD", "Median", "Range"), digit = 0)
tab1
}
