% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfrtf2df.r
\name{qc_rtf2df}
\alias{qc_rtf2df}
\title{Convert rtf to dataframe}
\usage{
qc_rtf2df(filename = "tsidem01", level = NULL)
}
\arguments{
\item{filename}{the rtf filename you want to read in}

\item{level}{Choose from "PDEV", "PREPROD" or "PROD". If left blank, find the rtf file in output folder on different levels in sequence of "PDEV", "PREPROD" and "PROD"}
}
\value{
a dataframe containing all the information from the rtf
}
\description{
read in rtf and convert it to dataframe
}
\examples{
dat <- qc_rtf2df("tsidem01", "PREPROD")
dat
}
