% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfshift.r
\name{qc_shift}
\alias{qc_shift}
\title{Create Shift table}
\usage{
qc_shift(
  input,
  acat = "ANRIND",
  bcat = "BNRIND",
  rowvar = c("PARAM", "AVISIT", "TRT01P"),
  drop_zero = TRUE
)
}
\arguments{
\item{input}{input dataframe}

\item{acat}{measurement variable}

\item{bcat}{baseline variable}

\item{rowvar}{row variable, Accepted values are a set of two variables such as 
\code{c("TRT01P", "AVISIT")} or three variables such as \code{c("PARAM", "TRT01P", "AVISIT")}}

\item{drop_zero}{if true, replace '0' with NA in N column}
}
\value{
shift table by rowvar
}
\description{
create shift table, accept up to 3 levels of row variables
}
\examples{
adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = TRUE))
  
param <- data.frame(PARAM = c("Test1", "Test2"))
visit <- data.frame(AVISIT = c("Baseline", "Visit1", "Visit2"))

adlb0 <- merge(adsl, param)
adlb <- merge(adlb0, visit)

adlb$ANRIND <- factor(sample(c("Low", "Medium", "High"), 60, replace = TRUE), 
                      levels = c("Low", "Medium", "High"))
adlb$BNRIND <- factor(sample(c("Low", "Medium", "High"), 60, replace = TRUE), 
                      levels = c("Low", "Medium", "High"))

tab1 <- qc_shift(adlb, "ANRIND", "BNRIND", rowvar = c("PARAM", "TRT01P", "AVISIT"))
tab1
}
