% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{num_row}
\alias{num_row}
\title{Create Rows for Continuous Variables}
\usage{
num_row(data, colvar, rowvar, digit)
}
\arguments{
\item{data}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable}

\item{digit}{number of decimal place to report}
}
\value{
dataframe with demographic rows
}
\description{
Function to create rows for continuous variables in demographic table
}
\examples{
num_row(adsl, "TRT01P", "AGE", digit = 1)
}
