% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_batchrun.r
\name{qc_batchrun}
\alias{qc_batchrun}
\title{Batch Run QC Scripts}
\usage{
qc_batchrun(files, path = ".")
}
\arguments{
\item{files}{a character vector of full/relative path names}

\item{path}{the folder path of html output file, in which a 'result' subfolder will be created to hold all compare results. The default corresponds to the working directory, getwd().}
}
\value{
A summary table of comparison results will display in Viewer
}
\description{
Execute R qc scripts end with \code{\link{qc_compare2xlsx}}
}
\examples{
\dontrun{
### select all r scripts start with 'qct' on SPACE
qc_files <- list.files(path = qc[["PDEV"]], pattern = "qct.*\\\\.r", 
                       full.names = TRUE)

qc_batchrun(files = qc_files, path = qc[["PDEV"]])
}
}
