% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_batchrun.r
\name{qc_batchrun}
\alias{qc_batchrun}
\title{Batch Run QC Scripts}
\usage{
qc_batchrun(files, path = NULL)
}
\arguments{
\item{files}{one or more filenames}

\item{path}{the path where batchrun result compare_results.html stores.}
}
\value{
A summary table of comparison results will display in Viewer
}
\description{
Execute R qc scripts end with \code{\link{qc_compare2xlsx}}
}
\examples{
\dontrun{
### select all r scripts start with 'qct' on SPACE
qc_files <- list.files(path = read_path(rptdrv, "qc"), pattern = "qct.*\\\\.r", 
                       full.names = TRUE)

qc_batchrun(qc_files)
}
}
