% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfbatchrun.r
\name{qc_batchrun}
\alias{qc_batchrun}
\title{Batch Run QC Scripts}
\usage{
qc_batchrun(files, path = qc)
}
\arguments{
\item{files}{one or more filenames}

\item{path}{the path of where temp1.txt is stored. Default is qc}
}
\value{
A summary table of comparison results will display in Viewer
}
\description{
Execute R qc scripts with \code{qc_compare2xlsx}
}
\examples{
\dontshow{if (exists("rptdrv")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
### select all r scripts start with 'qct' on SPACE
qc_files <- list.files(path = read_path(rptdrv, "qc"), pattern = "qct.*\\\\.r", 
                       full.names = TRUE)

qc_batchrun(qc_files)
\dontshow{\}) # examplesIf}
}
