% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_batchrun.r
\name{qc_batchrun}
\alias{qc_batchrun}
\title{Batch Run QC Scripts}
\usage{
qc_batchrun(files, path = ".", parallel = FALSE)
}
\arguments{
\item{files}{a character vector of full/relative path names}

\item{path}{the folder path of HTML output file. By default the path is the working directory, getwd().}

\item{parallel}{whether to batch run in parallel, default is FALSE}
}
\value{
A summary table of comparison results will display in html
}
\description{
Execute R qc scripts by batch with log
}
\examples{
\dontrun{
### select all r scripts start with 'qct' on SPACE
qc_files <- list.files(path = qc[["PDEV"]], pattern = "qct.*\\\\.r", 
                       full.names = TRUE)

qc_batchrun(files = qc_files, path = qc[["PDEV"]], parallel = TRUE)
}
}
