% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfdemofuns.r
\name{qc_cat_row}
\alias{qc_cat_row}
\title{Create Rows for Categorical Variables}
\usage{
qc_cat_row(input, colvar = "TRT01P", rowvar = "SEX", N_row, keep = TRUE)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{rowvar}{row variable}

\item{N_row}{a dataframe with denominator N}

\item{keep}{if = TRUE, keep all factor levels}
}
\value{
dataframe with demographic rows
}
\description{
Function to create rows for categorical variables in demographic table
}
\examples{
adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = T),
  SEX = as.factor(sample(c("Female", "Male"), 10, replace = T)))

### Create analysis row first
first_row <- qc_cntrow1(input = adsl, colvar = "TRT01P", row_text = "Analysis set: Safety")
  
tab1 <- qc_cat_row(adsl, "TRT01P", "SEX", N_row = first_row[[1]])
tab1
}
