% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountrow1.r
\name{qc_cntrow1}
\alias{qc_cntrow1}
\title{Create Analysis Set Row}
\usage{
qc_cntrow1(
  input,
  colvar = "TRT01P",
  row_text = "Analysis set: Safety",
  subset = "TRUE"
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{row_text}{row text}

\item{subset}{subset criteria. Default = "TRUE" means no subsetting}
}
\value{
Analysis set row list containing two elements.
The first element is a dataframe with N that can be passed to following functions
The second element is the analysis set row that can be combined with other following rows
}
\description{
Function to create analysis set row
}
\examples{
adae <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = T))
  
first_row <- qc_cntrow1(input = adae, colvar = "TRT01P", row_text = "Analysis set: Safety")

first_row$N_row
first_row$first_row
}
