% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcountrow1_chg.r
\name{qc_cntrow1_chg}
\alias{qc_cntrow1_chg}
\title{Create Analysis Set Row for Change From Baseline Table}
\usage{
qc_cntrow1_chg(
  input,
  colvar = "TRT01P",
  row_text = "Analysis set: Full",
  subset = NULL
)
}
\arguments{
\item{input}{input dataframe}

\item{colvar}{column variable}

\item{row_text}{row text}

\item{subset}{subset criteria. Default = NULL means no subsetting}
}
\value{
Analysis set row for change from baseline table
}
\description{
Function to create analysis set row for change from baseline table
}
\examples{
adlb <- data.frame(
  USUBJID = 1:10,
  TRT01P = factor(sample(c("A", "B", "C"), 10, replace = TRUE)))
  
first_row <- qc_cntrow1_chg(input = adlb, colvar = "TRT01P", row_text = "Analysis set: Full")
first_row
}
