% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfcompare2xlsx.r
\name{qc_compare2xlsx}
\alias{qc_compare2xlsx}
\title{Compare dataframes}
\usage{
qc_compare2xlsx(
  qc,
  rtf,
  path = qc,
  filename = "TSIDEM01",
  max_diff = 50,
  max_diff_per_var = 10,
  version = TRUE
)
}
\arguments{
\item{qc}{the dataframe you create in QC}

\item{rtf}{the dataframe read in from rtf}

\item{path}{the path to put xlsx format compare result}

\item{filename}{tableid}

\item{max_diff}{maximum number of differences to report}

\item{max_diff_per_var}{maximum number of differences per variable to report}

\item{version}{add version control by current date. Default = TURE}
}
\value{
compare result in xlsx
}
\description{
compare two dataframe and output the result to xlsx
}
\examples{
qc_compare2xlsx(qc = tab_qc, rtf = tab_rtf, path = qc, filename = "tableid")
}
