% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rtf2df.r
\name{qc_rtf2df}
\alias{qc_rtf2df}
\title{Convert RTF files to Data frame}
\usage{
qc_rtf2df(filename = "tsidem01", path = NULL)
}
\arguments{
\item{filename}{the RTF file name you want to read in}

\item{path}{the path of RTF file}
}
\value{
a data frame containing all data information from the RTF excluding titles and footnotes
}
\description{
Read in RTF files and convert it to data frame
}
\examples{
\dontrun{
dat <- qc_rtf2df("tsidem01", "PREPROD")
dat
}
}
