% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfshift.r
\name{qc_shift}
\alias{qc_shift}
\title{Create Shift table}
\usage{
qc_shift(
  input,
  acat = "ANRIND",
  bcat = "BNRIND",
  rowvar = c("PARAM", "AVISIT", "TRT01P"),
  drop_zero = TRUE
)
}
\arguments{
\item{input}{input dataframe}

\item{acat}{measurement variable}

\item{rowvar}{row variable (can be a set of two variables such as c("TRT01P", "AVISIT") or three variables such as c("PARAM", "TRT01P", "AVISIT"))}

\item{chg}{change from baseline measurement variable}

\item{stats_list}{stats variables to display. Accepted values are c("N", "Mean", "SD", "Median", "Min", "Max", "CV", "Base_mean") and the order of variables matters}

\item{max_digit}{largest number of digit to display}

\item{keep}{if = TRUE, keep all levels. Default = FALSE}
}
\value{
shift table by rowvar
}
\description{
create shift table, accept up to 3 levels of row variables
}
\examples{
adsl <- data.frame(
  USUBJID = 1:10,
  TRT01P = sample(c("A", "B", "C"), 10, replace = T))
  
param <- c("Test1", "Test2")
visit <- c("Baseline", "Visit1", "Visit2")

adlb0 <- merge(adsl, param)
adlb <- merge(adlb0, visit)

adlb$ANRIND <- factor(sample(c("Low", "Medium", "High"), 60), levels = c("Low", "Medium", "High"))
adlb$BNRIND <- factor(sample(c("Low", "Medium", "High"), 60), levels = c("Low", "Medium", "High"))

tab1 <- qc_shift(adlb, "ANRIND", "BNRIND", rowvar = c("PARAM", "TRT01P", "AVISIT"))
tab1
}
