% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcfround_sas.r
\name{round_sas}
\alias{round_sas}
\title{Round a numeric vector; halves will be rounded up as in SAS.}
\usage{
round_sas(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector to round.}

\item{digits}{how many digits should be displayed after the decimal point?}
}
\description{
Please see \code{janitor::round_half_up}
}
\examples{
round_sas(12.5)
round_sas(1.125, 2)
round_sas(1.125, 1)
round_sas(-0.5, 0) # negatives get rounded away from zero

}
